+++ src/SHBias.f95
@@ -50,10 +50,10 @@ subroutine SHBias(Shh, lwin, incspectra, ldata, outcspectra, save_cg, &
 
     real(dp), intent(in) :: Shh(:), incspectra(:)
     real(dp), intent(out) :: outcspectra(:)
-    integer, intent(in) :: lwin, ldata
-    integer, intent(in), optional :: save_cg
-    integer, intent(out), optional :: exitstatus
-    integer :: l, i, j, lmax, imin, imax, astat
+    integer(int32), intent(in) :: lwin, ldata
+    integer(int32), intent(in), optional :: save_cg
+    integer(int32), intent(out), optional :: exitstatus
+    integer(int32) :: l, i, j, lmax, imin, imax, astat
     real(dp) :: wig(2*lwin+ldata+1)
     real(dp), allocatable, save :: cg2(:,:,:)
